% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate an Object in a Cycle}
\usage{
rotate(x, n, ...)
}
\arguments{
\item{x}{object to be rotated.}

\item{n}{number of positions to rotate (left); negative to rotate right.}

\item{...}{passed to methods.}
}
\value{
a new object with elements shifted left by `n` positions
}
\description{
Shift the elements of a an object to the left, cycling elements to the
end of the object as they are removed from the start. Negative values
shift to the right.
}
\examples{
rotate(LETTERS[1:5])
rotate(LETTERS[1:5], 2)
rotate(LETTERS[1:5], -1)
}
